requirements.txt
robotframework
robotframework-seleniumlibrary
robotframework-pageobjectlibrary
robotframework-debuglibrary>=2
pyyaml